feat: support sub-1 FPS for video frame extraction (e.g., 0.5 FPS = 1 frame/2s)

+++ app/api/endpoints.py
@@ -189,7 +189,7 @@ async def predict(
     reparam_sigma_b: float = Form(0.0),
     text_embed_type_a: str = Form("projected"),
     text_embed_type_b: str = Form("projected"),
-    video_fps: int = Form(1),
+    video_fps: float = Form(1.0),
     # DATASET:QVHighlights specific fields
     qv_query: Optional[str] = Form(None),
     qv_vid: Optional[str] = Form(None),
